requirements.txt
flair
pymediawiki
transformers
pytorch
discord==1.7.3
python-dotenv~=0.19.0
